+++ src/pre_init.ts
@@ -1,4 +1,4 @@
-let serverURL = location.origin;
+let serverURL = (location.origin == "https://codeotter.dev" ? location.origin : "http://localhost:3000");
 // let serverURL = "http://localhost:3000";
 // let serverURL = "http://claebcode.top:3000";
 // let serverURL = "http://157.245.10.20:3000" // DigitalOcean URL